feat(material): emit laser fluence instead of PWM power fraction
The burn emitter now converts the 0-255 PWM power map into a float32
fluence map (J/cm2) via the new LaserPhysics struct (wavelength, optical
wattage, spot size, scan speed), threaded through AssembleCtx and the
ComputePayload. The outline burn emitter scales by the step's power
fraction so low-power vector cuts no longer render full char, and
RasterEffect.power becomes RasterEffect.fluence. MaterialFoldSpec and
MaterialState carry wavelength/power for the renderer's absorption
lookup, the fold max-reduces fluence into an F32 surface_map, and
CompressedArray gains from_vec_f32_with_shape/from_float32_2d.

+++ tests/ops/material/test_ops_material_fold_raster.py
@@ -27,7 +27,7 @@ def _fold(entries, grid=None):
 
 def _raster(values, origin_mm=(0.0, 0.0), px_per_mm=(1.0, 1.0)):
     return RasterEffect(
-        np.asarray(values, dtype=np.uint8),
+        np.asarray(values, dtype=np.float32),
         origin_mm=origin_mm,
         px_per_mm=px_per_mm,
     )
